ci: Add clippy and rustfmt checks, fix all lint warnings (#9)
## Description

- Add a `lint` CI job that runs `cargo fmt -- --check` and `cargo clippy
-- -D warnings`
- Fix all clippy warnings across the codebase (or disable in Cargo.toml)
- Move CI dependency installation to a script so it can be called from
multiple jobs

## Tests

CI

+++ src/ppu.rs
@@ -193,9 +193,9 @@ impl CgbColor {
 
     pub fn to_color32(&self) -> Color32 {
         Color32::from_rgb(
-            map_5_bit_color_to_8_bit(self.red() as u8),
-            map_5_bit_color_to_8_bit(self.green() as u8),
-            map_5_bit_color_to_8_bit(self.blue() as u8),
+            map_5_bit_color_to_8_bit(self.red()),
+            map_5_bit_color_to_8_bit(self.green()),
+            map_5_bit_color_to_8_bit(self.blue()),
         )
     }
 }
